// ArchiCore/src/app/api-client/model/dossierCreateRequest.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface DossierCreateRequest { 
    codeDossier: string;
    raisonSociale: string;
    typeEntreprise?: string;
    adresse?: string;
    phone?: string;
    email?: string;
    bp?: string;
}