// ArchiCore/src/app/api-client/model/dossierUpdateRequest.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface DossierUpdateRequest { 
    raisonSociale: string;
    typeEntreprise?: string;
    adresse?: string;
    phone?: string;
    email?: string;
    bp?: string;
}